+++ adk.h
@@ -104,7 +104,7 @@ public:
                 return 0;
         };
 
-        virtual uint8_t GetAddress() {
+        virtual uint32_t GetAddress() {
                 return bAddress;
         };
